requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0 
passlib[bcrypt]==1.7.4                 
python-multipart==0.0.9               
pymongo==4.7.2
motor==3.4.0
beanie==1.30.0
python-dotenv==1.0.1                 
pydantic-settings==2.2.1         
langchain
langgraph
langchain_openai
tiktoken==0.7.0
redis==5.0.4
httpx==0.27.0
websockets==12.0                         
